1. 抽象硬件层相关接口封装 2. 清除编译告警 3. 调整at_cmd与at_write中锁的使用
