+++ drivers/devices/nb/bc95.c
@@ -22,6 +22,7 @@ int32_t nb_data_handler(void* arg,int8_t * buf, int32_t len)
         return -1;
     }
     AT_LOG("cmd in");
+	return 0;
 }
 
 static int neul_bc95_hex_to_str(const char *bufin, int len, char *bufout)
@@ -44,7 +45,7 @@ static int neul_bc95_hex_to_str(const char *bufin, int len, char *bufout)
     return 0; 
 }
 
-static int neul_bc95_str_to_hex(const unsigned char *bufin, int len, char *bufout)
+static int neul_bc95_str_to_hex(const char *bufin, int len, char *bufout)
 {
     int i = 0;
     if (NULL == bufin || len <= 0 || NULL == bufout)
@@ -97,7 +98,7 @@ int nb_get_msg(void)
 		}
 		msgcnt--;
 	}
-
+	return 0;
 }
 
 int32_t nb_send_test(void)
@@ -128,7 +129,7 @@ int32_t nb_set_cdpserver(const char *ipaddr)
 		return ret;
 	}
 	ret = at.cmd((int8_t*)cmd2, strlen(cmd2), ipaddr, NULL);
-	osDelay(1000);
+	LOS_TaskDelay(1000);
 	ret = at.cmd((int8_t*)cmdNNMI, strlen(cmdNNMI), "OK", NULL);
 	//ret = at.cmd((int8_t*)cmdCGP, strlen(cmdCGP), NULL, NULL);
 	return ret;
@@ -149,7 +150,7 @@ int32_t nb_send_coap_payload(int32_t id ,const uint8_t *buf, uint32_t len)
 	char* str = NULL;
 	int curcnt = 0;
 	static int sndcnt = 0;
-	neul_bc95_hex_to_str((unsigned char*)buf, len, tmpbuf);
+	neul_bc95_hex_to_str((const char*)buf, len, tmpbuf);
 	sprintf(cmd, "%s%d,%s%c",cmd1,strlen(tmpbuf),tmpbuf,'\r');
 	ret = at.cmd((int8_t*)cmd, strlen(cmd), "OK", NULL);
 	if(ret < 0)
@@ -169,7 +170,6 @@ int neul_bc95_udp_read(int socket,char *buf, int maxrlen, int mode)
 {
     //AT+NSORF=0,4 
     char *cmd = "AT+NSORF=";
-    int ret = 0;
     char *str = "AT+NSORF=0,4";
     int rlen = 2;
     int rskt = -1;
@@ -186,12 +186,12 @@ int neul_bc95_udp_read(int socket,char *buf, int maxrlen, int mode)
 	//sscanf(rbuf,"\r+NSONMI:0,%d\r",&rlen);
 		
     sprintf(wbuf, "%s%d,%d\r", cmd, socket, maxrlen);
-    ret = at.cmd((int8_t*)str, strlen(str), "OK", rbuf);;
+    at.cmd((int8_t*)str, strlen(str), "OK", rbuf);
 
     sscanf(rbuf, "\r%d,%s,%d,%d,%s,%d\r%s", &rskt,tmpbuf,&port,&rlen,tmpbuf+22,&readleft,wbuf);
     if (rlen>0)
     {
-        neul_bc95_str_to_hex((const unsigned char *)(tmpbuf+22),rlen*2, buf);
+        neul_bc95_str_to_hex((const char *)(tmpbuf+22),rlen*2, buf);
     }
 	
     return rlen;
@@ -219,15 +219,13 @@ int32_t nb_get_netstat(void)
 
 int32_t nb_create_udpsock(const int8_t * host, const int8_t *port, int32_t proto)
 {
-	int ret = 0;
-	int timecnt = 0;
 	int socket;
 	char *cmd = "AT+NSOCR=DGRAM,17,";//udp
 	char wbuf[1024]={0};
 	char rbuf[1024]={0};
 	//nb_set_cdpserver(host,strlen(host));
 
-    sprintf(wbuf, "%s%d,1\r", cmd, port);
+    sprintf(wbuf, "%s%d,1\r", cmd, (int)port);
 	at.cmd((int8_t*)wbuf, strlen(wbuf), "OK", rbuf);
 	sscanf(rbuf, "%d\r%s",&socket, tmpbuf);
 	//neul_bc95_hex_to_str(tmpbuf, readlen*2, coapmsg);
@@ -242,8 +240,6 @@ int32_t nb_create_udpsock(const int8_t * host, const int8_t *port, int32_t proto
 int32_t nb_close_udpsock(int socket)
 {
     char *cmd = "AT+NSOCL=";
-    int ret = 0;
-    char *str = NULL;
 
     memset(wbuf, 0, 1064);
     sprintf(wbuf, "%s%d\r", cmd, socket);
@@ -262,7 +258,6 @@ int32_t nb_init(void)
 {
 	char* ipaddr = CLOUD_IP;
 	int ret;
-	char repbuf[20]={0};
     at.init();
 	at.oob_register("+NNMI",strlen("+NNMI"), nb_data_handler);
 
@@ -270,12 +265,12 @@ int32_t nb_init(void)
 	while(1)
 	{
 		ret = nb_hw_detect();
-		if(ret == 1)
+		if(ret != AT_FAILED)
 			break;
-		osDelay(1000);
+		LOS_TaskDelay(1000);
 	}
 	nb_set_cdpserver(ipaddr);//
-	osDelay(3000);
+	LOS_TaskDelay(3000);
 	//nb_get_auto_connect();
 	//nb_connect(NULL, NULL, NULL);
     return AT_OK;
@@ -289,20 +284,20 @@ int32_t nb_connect(const int8_t * host, const int8_t *port, int32_t proto)
 	{
 		ret = nb_get_netstat();
 		//nb_check_csq();
-		if(ret == 1)
+		if(ret != AT_FAILED)
 		{
 			ret = nb_query_ip();
 			break;
 		}
-		osDelay(1000);
+		LOS_TaskDelay(1000);
 		timecnt++;
 	}
-	if(ret == 1)
+	if(ret != AT_FAILED)
 	{
 		ret = nb_query_ip();
 	}
 	//nb_set_cdpserver((const char *)host);
-	ret = nb_create_udpsock(NULL, port, NULL);
+	ret = nb_create_udpsock(NULL, port, 0);
     return ret;
 
 }
@@ -310,12 +305,12 @@ int32_t nb_connect(const int8_t * host, const int8_t *port, int32_t proto)
 int32_t nb_send(int32_t id , const uint8_t  *buf, uint32_t len)
 {
 	char *cmd = "AT+NSOST=";
-	char *str = "AT+NSOST=0,192.53.100.53,5683,1,11\r";
+	//char *str = "AT+NSOST=0,192.53.100.53,5683,1,11\r";
 	
 	memset(wbuf, 0, 1064);
 	memset(tmpbuf, 0, 1064);
 	neul_bc95_str_to_hex((const char *)buf, len, tmpbuf);
-	sprintf(wbuf, "%s%d,%s,%d,%d,%s\r",cmd,id,"192.53.100.53",5683,len,tmpbuf);//这些参数需要socket结构体保存
+	sprintf(wbuf, "%s%d,%s,%d,%d,%s\r",cmd,(int)id,"192.53.100.53",(int)5683,(int)len,tmpbuf);//这些参数需要socket结构体保存
 	return at.cmd((int8_t*)wbuf, strlen(wbuf), "OK", NULL);
 }
 
@@ -333,7 +328,7 @@ int32_t nb_close(int32_t socket)
 {
     char *cmd = "AT+NSOCL=";
 	memset(wbuf, 0, 1064);
-	sprintf(wbuf, "%s%d\r", cmd, socket);
+	sprintf(wbuf, "%s%d\r", cmd, (int)socket);
 	return at.cmd((int8_t*)wbuf, strlen(wbuf), "OK", NULL);
 	//全局存储remote信息
 }
@@ -348,9 +343,20 @@ int32_t nb_deinit(void)
     return nb_reset();
 }
 
+at_config at_user_conf = {
+    .name = AT_MODU_NAME,
+    .usart = USART3,
+    .buardrate = AT_BUARDRATE,
+    .irqn = AT_USART_IRQn,
+    .linkid_num = AT_MAX_LINK_NUM,
+    .user_buf_len = MAX_AT_USERDATA_LEN,
+    .cmd_begin = AT_CMD_BEGIN,
+    .line_end = AT_LINE_END,
+    .mux_mode = 1, //support multi connection mode
+    .timeout = AT_CMD_TIMEOUT,   //  ms
+};
+
 at_adaptor_api at_interface = {
-    .name = (int8_t*)AT_MODU_NAME,
-    .timeout = 2000,   // 2000 tick
     .init = nb_init,    
 
     .connect = nb_connect,
