+++ drivers/devices/wifi/esp8266.c
@@ -282,21 +282,7 @@ int32_t esp8266_deinit(void)
 
 int32_t esp8266_init()
 {
-    at_config at_user_conf = {
-        .name = AT_MODU_NAME,
-        .usart = USART3,
-        .buardrate = AT_BUARDRATE,
-        .irqn = AT_USART_IRQn,
-        .linkid_num = AT_MAX_LINK_NUM,
-        .user_buf_len = MAX_AT_USERDATA_LEN,
-        .cmd_begin = AT_CMD_BEGIN,
-        .line_end = AT_LINE_END,
-        .mux_mode = 1, //support multi connection mode
-        .timeout = AT_CMD_TIMEOUT,   //  ms
-
-    };
-    at.init(&at_user_conf);
-    
+    at.init();
     //at.add_listener((int8_t*)AT_DATAF_PREFIX, NULL, esp8266_data_handler);
     at.oob_register(AT_DATAF_PREFIX, strlen(AT_DATAF_PREFIX), esp8266_data_handler);
 #ifdef 	USE_USARTRX_DMA
@@ -320,7 +306,19 @@ int32_t esp8266_init()
     AT_LOG("get ip:%s, gw:%s mac:%s", ip, gw, mac);
     return AT_OK;
 }
+at_config at_user_conf = {
+    .name = AT_MODU_NAME,
+    .usart = USART3,
+    .buardrate = AT_BUARDRATE,
+    .irqn = AT_USART_IRQn,
+    .linkid_num = AT_MAX_LINK_NUM,
+    .user_buf_len = MAX_AT_USERDATA_LEN,
+    .cmd_begin = AT_CMD_BEGIN,
+    .line_end = AT_LINE_END,
+    .mux_mode = 1, //support multi connection mode
+    .timeout = AT_CMD_TIMEOUT,   //  ms
 
+};
 
 at_adaptor_api at_interface = {
